restructure unbalanced.py
